requirements.txt
fastapi
uvicorn[standard]
playwright
sentence-transformers
qdrant-client
pydantic
apscheduler
python-dotenv
requests